requirements.txt
Flask
python-dotenv
requests
spotipy
transformers
torch
Pillow
comet-ml
torchvision